+++ src/Internal/WhatWG.php
@@ -931,11 +931,11 @@ class WhatWG {
 	 * The "XML serialization" algorithm.
 	 * @see https://w3c.github.io/DOM-Parsing/#dfn-xml-serialization
 	 * @param Node $node
-	 * @param bool $requireWellFormed
+	 * @param array $options
 	 * @param string[] &$markup
 	 */
 	public static function xmlSerialize(
-		Node $node, bool $requireWellFormed, array &$markup
+		Node $node, array $options, array &$markup
 	): void {
 		$contextNamespace = null;
 		$prefixMap = new NamespacePrefixMap();
@@ -944,7 +944,7 @@ class WhatWG {
 		try {
 				$node->_xmlSerialize(
 					$contextNamespace, $prefixMap, $prefixIndex,
-					$requireWellFormed, $markup
+					$options, $markup
 				);
 		} catch ( \Throwable $t ) {
 			Util::error( 'InvalidStateError' );
@@ -959,15 +959,15 @@ class WhatWG {
 	 * @param ?string $namespace
 	 * @param NamespacePrefixMap $prefixMap
 	 * @param int &$prefixIndex
-	 * @param bool $requireWellFormed
+	 * @param array $options
 	 * @param string[] &$markup accumulator for the result
 	 */
 	public static function xmlSerializeElement(
 		Element $el, ?string $namespace,
 		NamespacePrefixMap $prefixMap, int &$prefixIndex,
-		bool $requireWellFormed, array &$markup
+		array $options, array &$markup
 	) {
-		if ( $requireWellFormed ) {
+		if ( $options['requireWellFormed'] ?? false ) {
 			if (
 				strpos( $el->getLocalName(), ':' ) !== false ||
 				!self::is_valid_xml_name( $el->getLocalName() )
@@ -1002,7 +1002,7 @@ class WhatWG {
 				$candidatePrefix = $map->retrievePreferredPrefix( $ns, $prefix );
 			}
 			if ( $prefix === 'xmlns' ) {
-				if ( $requireWellFormed ) {
+				if ( $options['requireWellFormed'] ?? false ) {
 					throw new BadXMLException();
 				}
 				$candidatePrefix = $prefix;
@@ -1029,7 +1029,7 @@ class WhatWG {
 				$qualifiedName .= $prefix . ':' . $el->getLocalName();
 				$markup[] = $qualifiedName;
 				$markup[] = ' xmlns:' . $prefix . '="';
-				self::xmlSerializeAttrValue( $ns, $requireWellFormed, $markup );
+				self::xmlSerializeAttrValue( $ns, $options, $markup );
 				$markup[] = '"';
 				if ( $localDefaultNamespace !== null ) {
 					$inheritedNs = $localDefaultNamespace;
@@ -1050,7 +1050,7 @@ class WhatWG {
 				$inheritedNs = $ns;
 				$markup[] = $qualifiedName;
 				$markup[] = ' xmlns="';
-				self::xmlSerializeAttrValue( $ns, $requireWellFormed, $markup );
+				self::xmlSerializeAttrValue( $ns, $options, $markup );
 				$markup[] = '"';
 			} else {
 				// The node has a local default namespace that matches ns
@@ -1062,21 +1062,23 @@ class WhatWG {
 		// Serialize the attributes
 		self::xmlSerializeAttributes(
 			$el->getAttributes(), $map, $prefixIndex, $localPrefixesMap,
-			$ignoreNamespaceDefinitionAttribute, $requireWellFormed,
+			$ignoreNamespaceDefinitionAttribute, $options,
 			$markup
 		);
 
 		if (
 			$ns === Util::NAMESPACE_HTML &&
 			( !$el->hasChildNodes() ) &&
-			( self::$emptyElements[$el->getLocalName()] ?? false )
+			( self::$emptyElements[$el->getLocalName()] ?? false ) &&
+			!( $options['noEmptyTag'] ?? false )
 		) {
 			$markup[] = ' />';
 			return;
 		}
 		if (
 			$ns !== Util::NAMESPACE_HTML &&
-			( !$el->hasChildNodes() )
+			( !$el->hasChildNodes() ) &&
+			!( $options['noEmptyTag'] ?? false )
 		) {
 			$markup[] = '/>';
 			return;
@@ -1091,14 +1093,14 @@ class WhatWG {
 				$el->getContent() :
 				$el->getOwnerDocument()->createDocumentFragment();
 			$templateContents->_xmlSerialize(
-				$inheritedNs, $map, $prefixIndex, $requireWellFormed,
+				$inheritedNs, $map, $prefixIndex, $options,
 				$markup
 			);
 		} else {
 			// handle element contents
 			for ( $child = $el->getFirstChild(); $child !== null; $child = $child->getNextSibling() ) {
 				$child->_xmlSerialize(
-					$inheritedNs, $map, $prefixIndex, $requireWellFormed,
+					$inheritedNs, $map, $prefixIndex, $options,
 					$markup
 				);
 			}
@@ -1115,18 +1117,18 @@ class WhatWG {
 	 * @param int &$prefixIndex
 	 * @param array<string,string> &$localPrefixesMap
 	 * @param bool $ignoreNamespaceDefinitionAttribute
-	 * @param bool $requireWellFormed
+	 * @param array $options
 	 * @param string[] &$markup accumulator for the result
 	 */
 	public static function xmlSerializeAttributes(
 		NamedNodeMap $attributes, NamespacePrefixMap $map, int &$prefixIndex,
 		array &$localPrefixesMap, bool $ignoreNamespaceDefinitionAttribute,
-		bool $requireWellFormed,
+		array $options,
 		array &$markup
 	) {
 		$localnameSet = [];
 		foreach ( $attributes as $attr ) {
-			if ( $requireWellFormed ) {
+			if ( $options['requireWellFormed'] ?? false ) {
 				$key = var_export(
 					[ $attr->getNamespaceURI(), $attr->getLocalName() ],
 					true
@@ -1164,7 +1166,7 @@ class WhatWG {
 							continue;
 						}
 					}
-					if ( $requireWellFormed ) {
+					if ( $options['requireWellFormed'] ?? false ) {
 						if ( $attr->getValue() === Util::NAMESPACE_XMLNS ) {
 							throw new BadXMLException();
 						}
@@ -1193,7 +1195,7 @@ class WhatWG {
 					);
 					$markup[] = ' xmlns:' . $candidatePrefix . '="';
 					self::xmlSerializeAttrValue(
-						$attrNs, $requireWellFormed, $markup
+						$attrNs, $options, $markup
 					);
 					$markup[] = '"';
 				}
@@ -1202,7 +1204,7 @@ class WhatWG {
 			if ( $candidatePrefix !== null ) {
 				$markup[] = $candidatePrefix . ':';
 			}
-			if ( $requireWellFormed ) {
+			if ( $options['requireWellFormed'] ?? false ) {
 				if (
 					strpos( $attr->getLocalName(), ':' ) !== false ||
 					( !self::is_valid_xml_name( $attr->getLocalName() ) ) ||
@@ -1214,7 +1216,7 @@ class WhatWG {
 			$markup[] = $attr->getLocalName();
 			$markup[] = '="';
 			self::xmlSerializeAttrValue(
-				$attr->getValue(), $requireWellFormed, $markup
+				$attr->getValue(), $options, $markup
 			);
 			$markup[] = '"';
 		}
@@ -1224,16 +1226,16 @@ class WhatWG {
 	 * Serialize an attribute value (for XML).
 	 * @see https://w3c.github.io/DOM-Parsing/#dfn-serializing-an-attribute-value
 	 * @param ?string $value
-	 * @param bool $requireWellFormed
+	 * @param array $options
 	 * @param string[] &$markup
 	 */
 	public static function xmlSerializeAttrValue(
-		?string $value, bool $requireWellFormed, array &$markup
+		?string $value, array $options, array &$markup
 	): void {
 		if ( $value === null ) {
 			return; // "The empty string"
 		}
-		if ( $requireWellFormed ) {
+		if ( $options['requireWellFormed'] ?? false ) {
 			if ( !self::is_valid_xml_chars( $value ) ) {
 				throw new BadXMLException();
 			}